+++ x/eibc/types/genesis_test.go
@@ -3,8 +3,10 @@ package types_test
 import (
 	"testing"
 
-	"github.com/dymensionxyz/dymension/v3/x/eibc/types"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/require"
+
+	"github.com/dymensionxyz/dymension/v3/x/eibc/types"
 )
 
 func TestGenesisState_Validate(t *testing.T) {
@@ -17,11 +19,33 @@ func TestGenesisState_Validate(t *testing.T) {
 			desc:     "default is valid",
 			genState: types.DefaultGenesis(),
 			valid:    true,
-		},
-		{
-			desc:     "valid genesis state",
-			genState: &types.GenesisState{},
-			valid:    true,
+		}, {
+			desc: "valid genesis state",
+			genState: &types.GenesisState{
+				Params:       validParams,
+				DemandOrders: []types.DemandOrder{validDemandOrder},
+			},
+			valid: true,
+		}, {
+			desc: "invalid params",
+			genState: &types.GenesisState{
+				Params: types.Params{
+					TimeoutFee: sdk.NewDec(-1),
+					ErrackFee:  sdk.NewDec(-1),
+				},
+			},
+			valid: false,
+		}, {
+			desc:     "invalid demand order",
+			genState: &types.GenesisState{DemandOrders: []types.DemandOrder{{}}, Params: types.DefaultParams()},
+			valid:    false,
+		}, {
+			desc: "duplicate demand order",
+			genState: &types.GenesisState{DemandOrders: []types.DemandOrder{
+				validDemandOrder,
+				validDemandOrder,
+			}, Params: types.DefaultParams()},
+			valid: false,
 		},
 	} {
 		t.Run(tc.desc, func(t *testing.T) {
@@ -33,4 +57,17 @@ func TestGenesisState_Validate(t *testing.T) {
 			}
 		})
 	}
+}
+
+var validDemandOrder = types.DemandOrder{
+	Id:        "1",
+	Price:     sdk.Coins{sdk.NewInt64Coin("denom", 2)},
+	Fee:       sdk.Coins{sdk.NewInt64Coin("denom", 1)},
+	Recipient: "cosmos18wvvwfmq77a6d8tza4h5sfuy2yj3jj88yqg82a",
+}
+
+var validParams = types.Params{
+	EpochIdentifier: "hour",
+	TimeoutFee:      sdk.NewDecWithPrec(1, 1),
+	ErrackFee:       sdk.NewDecWithPrec(1, 1),
 }